Re-organized project inheritance
I really thought about how the inheritance hierarchy should be done for the game that I want, and my solution required me to completely redo some game functions such as damage and health. I realized this change was very desirable when writing the spawner script.

As a part of this change, I made the following term definitions:
Damager = An object that may deal damage. All damagers exist for a specified time.
Hazard = A damager that deals damage to the player specifically.
Enemy = A hazard that may take damage (has health)
Shooter = An enemy that may shoot projectiles

Attack = A damager that is produced by another object. Attacks save a reference to their sender.
Projectile = An attack that detaches itself from the player and flies off somewhere.
Bullet = A type of projectile that is destroyed upon hitting anything that is not its owner or another projectile.

Some of relationships between the objects above are not fully expressed in the code. (See the various behavior scripts) This is because I didn't have a use for expressing those relationships yet.

I also differentiated invincibility after getting hit from the 'damaged' effect (turning a little more red when getting hit). The damaged concept was applied to the turret enemy as well.

Features for the Future:
- More enemies
  - Enemies that just fly at the player
  - Hazards that extend from the walls
  - etc.
- Animations
  - Enemies appearing and disappearing
  - Enemy death
  - Bullet explosion
- Make enemies drop things
  - Coins for purchasing classes and upgrades to classes
  - Small damage upgrades

this took me all day lol