Changed some colors
Created my own extensive shape sprites that are to be used to build everything in the game. In addition to this, I put a lot more thought into the colors. I will have to change the colors of more things, change the shapes, and otherwise continue to overhaul the look of the game.